[core]Modularize notEqual.
